A lot of fucking changes. Please... next time when someone is going to start designing in WPF... remember: NEVER USE THE FUCKING DRAG & DROP!!! If you are going to make a mess in XAML code... KILL YOURSELF!!! (but to be clear - I appreciate the contribution to the project).
Prepared for testing (added some "TODO" comments which u can conveniently find using View->Task List window in VS. NOT TESTED!

It should be considered if we should separate the rest of bindings of master and slave sub-views in order to have one app which can handle master and slave simultaneously. Another option (more noob but I'm not against it) is to launch instance of our application each time we need another master or slave.

The end.